requirements.txt
discord-py-interactions
dotenv
